requirements.txt
boto3>=1.26.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
xgboost>=1.7.0
pyyaml>=6.0
pytest>=7.0.0
matplotlib>=3.6.0
seaborn>=0.12.0
jupyter>=1.0.0
